# feat: optimise BinningData, remove dead code