--- a/Tests/Benchmarks/CMakeLists.txt
+++ b/Tests/Benchmarks/CMakeLists.txt
@@ -22,6 +22,7 @@ endmacro()
 
 add_benchmark(AtlasStepper AtlasStepperBenchmark.cpp)
 add_benchmark(BoundaryCheck BoundaryCheckBenchmark.cpp)
+add_benchmark(BinUtility BinUtilityBenchmark.cpp)
 add_benchmark(EigenStepper EigenStepperBenchmark.cpp)
 add_benchmark(SolenoidField SolenoidFieldBenchmark.cpp)
 add_benchmark(SurfaceIntersection SurfaceIntersectionBenchmark.cpp)