@paulgessinger 2021-01-21
This changes the public API, but we probably wouldn't consider this breaking, right @asalzburger ?